refactor: move reused components to a common folder
This makes them easier to find.
